refactor(Utils): move price stuff to a dedicated file
