+++ src/components/ProofReceiptOnlineDeliveryCostsChip.vue
@@ -6,7 +6,7 @@
 
 <script>
 import constants from '../constants'
-import utils from '../utils.js'
+import price_utils from '../utils/price.js'
 
 export default {
   props: {
@@ -26,7 +26,7 @@ export default {
   },
   methods: {
     getPriceValue(priceValue, priceCurrency) {
-      return utils.prettyPrice(priceValue, priceCurrency)
+      return price_utils.prettyPrice(priceValue, priceCurrency)
     },
     getPriceValueDisplay(price) {
       price = parseFloat(price)
