+++ src/utils.js
@@ -66,25 +66,6 @@ function removeObjectFromArray(arr, obj) {
 }
 
 /**
- * fr: 4,00 €
- * en-US: €4.00
- */
-function prettyPrice(price, currency) {
-  return price.toLocaleString(navigator.language, {
-    style: 'currency',
-    currency: currency,
-    minimumFractionDigits: 2,
-    maximumFractionDigits: 2
-  })
-}
-
-function priceSum(priceList) {
-  return priceList.reduce((acc, price) => {
-    return acc + parseFloat(price.price) * (price.receipt_quantity ? parseFloat(price.receipt_quantity) : 1)
-  }, 0)
-}
-
-/**
  * OFF auth token format: 'username__uuid'
  */
 function getOFFUsernameFromAuthToken(token) {
@@ -129,10 +110,6 @@ function getLocaleLabelTagName(locale, labelId) {
   })
 }
 
-function getPriceTypeIcon(priceType) {
-  return constants[`PRICE_TYPE_${priceType}_ICON`] || constants.PRICE_ICON
-}
-
 function getProofTypeIcon(proofType) {
   return constants[`PROOF_TYPE_${proofType}_ICON`] || constants.PROOF_ICON
 }
@@ -145,8 +122,6 @@ export default {
   getURLOrigin,
   addObjectToArray,
   removeObjectFromArray,
-  prettyPrice,
-  priceSum,
   getOFFUsernameFromAuthToken,
   getCategoryName,
   getLocaleCategoryTags,
@@ -155,6 +130,5 @@ export default {
   getLocaleOriginTags,
   getLocaleLabelTags,
   getLocaleLabelTagName,
-  getPriceTypeIcon,
   getProofTypeIcon,
 }
